Moved tools lib to src/lib
