+++ src/lib/actions/describe-state-machine.js
@@ -1,4 +1,4 @@
-const { isValidArn } = require('./tools/validate');
+const { isValidArn } = require('../tools/validate');
 const { errors } = require('../../constants');
 
 function describeStateMachine(params, stateMachines) {
