+++ src/lib/actions/stop-execution.js
@@ -1,4 +1,4 @@
-const { isValidArn } = require('./tools/validate');
+const { isValidArn } = require('../tools/validate');
 const { errors, status } = require('../../constants');
 
 function stopExecution(params, executions) {
