+++ src/lib/tools/validate.js
